+++ .delivery/build_cookbook/attributes/default.rb
@@ -1,1 +1,0 @@
-default['chef']['server_url'] = 'https://azl-chef-srvr-01/organizations/trek'