+++ .delivery/build_cookbook/recipes/default.rb
@@ -4,6 +4,3 @@
 #
 # Copyright:: 2017, The Authors, All Rights Reserved.
 include_recipe 'delivery-truck::default'
-
-node.default['terraform']['version'] = '0.11.7'
-include_recipe 'terraform'