+++ metadata.rb
@@ -4,13 +4,14 @@ maintainer_email 'ray_crawford@trekbikes.com'
 license 'All Rights Reserved'
 description 'Installs/Configures cb_dvo_logging'
 long_description 'Installs/Configures cb_dvo_logging'
-version '2.2.24'
+version '2.2.25'
 
 source_url 'https://bitbucket.org/trekbikes/cb_dvo_logging'
 issues_url 'https://bitbucket.org/trekbikes/cb_dvo_logging/issues?status=new&status=open'
 
 chef_version '>= 12.1' if respond_to?(:chef_version)
 supports 'centos'
+supports 'windows'
 
 depends 'sumologic-collector', '~> 1.2.23'
 depends 'cb_dvo_addstorage'